+++ lib/src/flutter/steps/when_tap_widget_step.dart
@@ -19,12 +19,15 @@ import 'package:gherkin/gherkin.dart';
 ///   `When I tap "controlKey" widget"`
 class WhenTapWidget extends When1WithWorld<String, FlutterWorld> {
   @override
-  RegExp get pattern => RegExp(
-      r"I tap the {string} [button|element|label|icon|field|text|widget]");
+  RegExp get pattern =>
+      RegExp(r"I tap the {string} (?:button|element|label|icon|field|text|widget)$");
 
   @override
   Future<void> executeStep(String key) async {
-    await FlutterDriverUtils.tap(world.driver, find.byValueKey(key),
-        timeout: timeout * .9);
+    await FlutterDriverUtils.tap(
+      world.driver,
+      find.byValueKey(key),
+      timeout: timeout * .9,
+    );
   }
 }