+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added retry logic to the Futter driver connect call to handle the seemingly random connection failures
 * Ensured `AttachScreenshotOnFailedStepHook` cannot throw an unhandled exception causing the test run to stop
 * Added new well known step `When I tap the back button` which finds and taps the default page back button
+* Updated Gherkin library version
 
 ## [1.1.7+2] - 07/01/2019
 * Increased the Flutter driver reconnection delay to try and overcome some driver to app connection issues on slower machines